+++ src/remotes/dzrpbuffer/cspectremote.ts
@@ -1,7 +1,6 @@
 import {DzrpTransportRemote} from './dzrptransportremote';
 import {CSpectType, Settings} from '../../settings/settings';
 import {WithSocket} from './transportsocketmixin';
-import {DZRP} from '../dzrp/dzrpremote';
 
 
 
@@ -19,15 +18,6 @@ export class CSpectRemote extends WithSocket(DzrpTransportRemote) {
 		super(settingsDzrpType);
 		// Init
 		this.supportsBreakOnInterrupt = false;
-	}
-
-
-	/** Returns the default unsupported commands for the CSpect remote.
-	 */
-	protected getDefaultUnsupportedCommands(): number[] {
-		return [
-			DZRP.CMD_SET_BREAKPOINTS, DZRP.CMD_RESTORE_MEM, DZRP.CMD_LOOPBACK, DZRP.CMD_ADD_WATCHPOINT, DZRP.CMD_REMOVE_WATCHPOINT, DZRP.CMD_READ_STATE, DZRP.CMD_WRITE_STATE
-		];
 	}